// src/interfaces/PasswordToggleOptions.ts
/**
 * Password show toggle options.
 */
export default interface PasswordToggleOptions {
  /**
   * CSS class to be applied to the button that shows the password. Default is "bi bi-eye fs-2".
   */
  showButtonClass?: string;

  /**
   * CSS class to be applied to the button that hides the password. Default is "bi bi-eye-slash fs-2".
   */
  hideButtonClass?: string;
}